Add wyrefi.service
We split all EFI operations into a separate command. This command is run
before the net command and after the boot command. This provides a few
benefits.

First, all EFI operations are consolidated into a single command.

Second, we can clear EFI and reboot before bringing up networking. This
benefit is particularly important. If we need to clear the EFI, which
networking configuration should we use? The default one might not work.
But the saved one also might not work. We solve this problem by clearing
EFI early and rebooting.